refactor(stores): consolidate store imports for cleaner code
This commit refactors the way store modules are imported across the application. Instead of importing each store individually from their respective paths, all stores are now imported from a centralized index file. This change improves code readability and maintainability by reducing duplication and simplifying import statements throughout the codebase.

No external functionality has been altered, but this restructuring sets the stage for easier management of store-related imports in future development work.
