+++ src/features/meal-form/components/MealFormAccordionItem.tsx
@@ -3,8 +3,7 @@ import { useDisclosure } from '@mantine/hooks'
 import { useEffect } from 'react'
 
 import { DialogSaved } from '../../../components/DialogSaved'
-import { useCurrentDateStore } from '../../../stores/currentDate'
-import { useMealRecordsStore } from '../../../stores/mealRecords'
+import { useCurrentDateStore, useMealRecordsStore } from '../../../stores'
 import { createStringFromDate } from '../../../utils'
 import { FormsType } from '../types'
 import { createFoodInitialValues, saveMealRecord } from '../utils'
